requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-chroma==0.0.10
langchain-ollama==0.0.1
chromadb==0.4.22
streamlit==1.29.0
playwright==1.40.0
beautifulsoup4==4.12.2
tqdm==4.66.1
python-dotenv==1.0.0
asyncio==3.4.3
typing-extensions==4.8.0
rich==13.7.0